refactor: ArrayValue is now a tree that defers conversion to ibis
