+++ bigframes/core/compile/__init__.py
@@ -12,9 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Helpers to join ArrayValue objects."""
+from bigframes.core.compile.compiled import CompiledArrayValue
+from bigframes.core.compile.compiler import ArrayValueCompiler
 
-from bigframes.core.joins.merge import merge
-from bigframes.core.joins.name_resolution import JOIN_NAME_REMAPPER, JoinNameRemapper
-
-__all__ = ("merge", "JoinNameRemapper", "JOIN_NAME_REMAPPER")
+__all__ = [
+    "ArrayValueCompiler",
+    "CompiledArrayValue",
+]
