+++ bigframes/core/window_spec.py
@@ -1,0 +1,35 @@
+# Copyright 2023 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from dataclasses import dataclass
+import typing
+
+import bigframes.core.ordering as orderings
+
+
+@dataclass(frozen=True)
+class WindowSpec:
+    """
+    Specifies a window over which aggregate and analytic function may be applied.
+    grouping_keys: set of column ids to group on
+    preceding: Number of preceding rows in the window
+    following: Number of preceding rows in the window
+    ordering: List of columns ids and ordering direction to override base ordering
+    """
+
+    grouping_keys: typing.Sequence[str] = tuple()
+    ordering: typing.Sequence[orderings.OrderingColumnReference] = tuple()
+    preceding: typing.Optional[int] = None
+    following: typing.Optional[int] = None
+    min_periods: int = 0
